// src/api/models/page-res-select-counselee-res.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { SelectCounseleeRes } from './select-counselee-res';

/**
 * 
 * @export
 * @interface PageResSelectCounseleeRes
 */
export interface PageResSelectCounseleeRes {
    /**
     * 
     * @type {Array<SelectCounseleeRes>}
     * @memberof PageResSelectCounseleeRes
     */
    'content'?: Array<SelectCounseleeRes>;
    /**
     * 
     * @type {number}
     * @memberof PageResSelectCounseleeRes
     */
    'page'?: number;
    /**
     * 
     * @type {number}
     * @memberof PageResSelectCounseleeRes
     */
    'size'?: number;
    /**
     * 
     * @type {number}
     * @memberof PageResSelectCounseleeRes
     */
    'totalElements'?: number;
    /**
     * 
     * @type {number}
     * @memberof PageResSelectCounseleeRes
     */
    'totalPages'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PageResSelectCounseleeRes
     */
    'hasNext'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof PageResSelectCounseleeRes
     */
    'hasPrevious'?: boolean;
}