// src/api/models/select-counselee-res.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface SelectCounseleeRes
 */
export interface SelectCounseleeRes {
    /**
     * 
     * @type {string}
     * @memberof SelectCounseleeRes
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof SelectCounseleeRes
     */
    'name'?: string;
    /**
     * 
     * @type {number}
     * @memberof SelectCounseleeRes
     */
    'age'?: number;
    /**
     * 
     * @type {string}
     * @memberof SelectCounseleeRes
     */
    'dateOfBirth'?: string;
    /**
     * 
     * @type {string}
     * @memberof SelectCounseleeRes
     */
    'phoneNumber'?: string;
    /**
     * 
     * @type {string}
     * @memberof SelectCounseleeRes
     */
    'gender'?: SelectCounseleeResGenderEnum;
    /**
     * 
     * @type {string}
     * @memberof SelectCounseleeRes
     */
    'address'?: string;
    /**
     * 
     * @type {string}
     * @memberof SelectCounseleeRes
     */
    'affiliatedWelfareInstitution'?: string;
    /**
     * 
     * @type {string}
     * @memberof SelectCounseleeRes
     */
    'healthInsuranceType'?: SelectCounseleeResHealthInsuranceTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof SelectCounseleeRes
     */
    'counselCount'?: number;
    /**
     * 
     * @type {string}
     * @memberof SelectCounseleeRes
     */
    'lastCounselDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof SelectCounseleeRes
     */
    'registrationDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof SelectCounseleeRes
     */
    'careManagerName'?: string;
    /**
     * 
     * @type {string}
     * @memberof SelectCounseleeRes
     */
    'note'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof SelectCounseleeRes
     */
    'disability'?: boolean;
}

export const SelectCounseleeResGenderEnum = {
    Male: 'MALE',
    Female: 'FEMALE',
    Else: 'ELSE'
} as const;

export type SelectCounseleeResGenderEnum = typeof SelectCounseleeResGenderEnum[keyof typeof SelectCounseleeResGenderEnum];
export const SelectCounseleeResHealthInsuranceTypeEnum = {
    HealthInsurance: 'HEALTH_INSURANCE',
    MedicalAid: 'MEDICAL_AID',
    VeteransBenefits: 'VETERANS_BENEFITS',
    NonCovered: 'NON_COVERED'
} as const;

export type SelectCounseleeResHealthInsuranceTypeEnum = typeof SelectCounseleeResHealthInsuranceTypeEnum[keyof typeof SelectCounseleeResHealthInsuranceTypeEnum];